backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
praw
langid
python-dotenv
